pyproject.toml
[project]
name = "fastapi-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pyjwt>=2.10.1",
    "ruff>=0.12.7",
]
